+++ admission_control/admission-control-service/Cargo.toml
@@ -24,6 +24,7 @@ executable-helpers = { path = "../../common/executable-helpers" }
 grpc_helpers = { path = "../../common/grpc_helpers" }
 logger = { path = "../../common/logger" }
 mempool = { path = "../../mempool" }
+mempool-shared-proto = { path = "../../mempool/mempool-shared-proto" }
 metrics = { path = "../../common/metrics" }
 storage_client = { path = "../../storage/storage_client" }
 types = { path = "../../types" }
