+++ admission_control/admission_control_proto/Cargo.toml
@@ -14,7 +14,7 @@ prost = "0.5.0"
 
 failure = { package = "failure_ext", path = "../../common/failure_ext" }
 logger = { path = "../../common/logger" }
-mempool = { path = "../../mempool" }
+mempool-shared-proto = { path = "../../mempool/mempool-shared-proto" }
 types = { path = "../../types" }
 
 [dev-dependencies]
