+++ admission_control/admission_control_proto/build.rs
@@ -7,7 +7,7 @@ fn main() {
     let includes = [
         "../../types/src/proto",
         "src/proto",
-        "../../mempool/src/proto/shared",
+        "../../mempool/mempool-shared-proto/src/proto",
     ];
 
     grpcio_compiler::prost_codegen::compile_protos(
