+++ mempool/src/core_mempool/mod.rs
@@ -6,11 +6,7 @@ mod mempool;
 mod transaction;
 mod transaction_store;
 
-pub use self::{
-    index::TxnPointer,
-    mempool::Mempool as CoreMempool,
-    transaction::{MempoolAddTransactionStatus, TimelineState},
-};
+pub use self::{index::TxnPointer, mempool::Mempool as CoreMempool, transaction::TimelineState};
 
 #[cfg(test)]
 mod unit_tests;
