+++ mempool/src/core_mempool/transaction.rs
@@ -1,9 +1,6 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::proto::shared::mempool_status::MempoolAddTransactionStatusCode;
-use failure::prelude::*;
-use std::convert::TryFrom;
 use std::time::Duration;
 use types::{account_address::AccountAddress, transaction::SignedTransaction};
 
@@ -52,49 +49,4 @@ pub enum TimelineState {
     // transaction will never be qualified for broadcasting
     // currently we don't broadcast transactions originated on other peers
     NonQualified,
-}
-
-/// Status of transaction insertion operation
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct MempoolAddTransactionStatus {
-    /// Status code of the transaction insertion operation
-    pub code: MempoolAddTransactionStatusCode,
-    /// Message to give more details about the transaction insertion operation
-    pub message: String,
-}
-
-impl MempoolAddTransactionStatus {
-    /// Create a new MempoolAddTransactionStatus
-    pub fn new(code: MempoolAddTransactionStatusCode, message: String) -> Self {
-        Self { code, message }
-    }
-}
-
-//***********************************
-// Decoding/Encoding to Protobuffers
-//***********************************
-impl TryFrom<crate::proto::shared::mempool_status::MempoolAddTransactionStatus>
-    for MempoolAddTransactionStatus
-{
-    type Error = Error;
-
-    fn try_from(
-        proto: crate::proto::shared::mempool_status::MempoolAddTransactionStatus,
-    ) -> Result<Self> {
-        Ok(MempoolAddTransactionStatus::new(
-            proto.code(),
-            proto.message,
-        ))
-    }
-}
-
-impl From<MempoolAddTransactionStatus>
-    for crate::proto::shared::mempool_status::MempoolAddTransactionStatus
-{
-    fn from(status: MempoolAddTransactionStatus) -> Self {
-        let mut mempool_add_transaction_status = Self::default();
-        mempool_add_transaction_status.message = status.message;
-        mempool_add_transaction_status.set_code(status.code);
-        mempool_add_transaction_status
-    }
 }
